backend/requirements.txt
appnope
asgiref
backcall
decorator
Django
django-cors-headers
djangorestframework
django-spa
dj_database_url
gunicorn
ipython
ipython-genutils
jedi
matplotlib-inline
numpy
pandas
parso
pexpect
pickleshare
prompt-toolkit
psycopg2-binary
ptyprocess
Pygments
python-dateutil
pytz
six
sqlparse
toml
traitlets
typing-extensions
wcwidth
